chore: Re-organise schemas into submodules (MPT-7140)
