+++ app/routers/employees.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 
 from app.api_clients import APIModifierClient, OptscaleAuthClient, OptscaleClient, UserDoesNotExist
 from app.auth.auth import check_operations_account
-from app.schemas import EmployeeCreate, EmployeeRead
+from app.schemas.employees import EmployeeCreate, EmployeeRead
 from app.utils import wrap_http_error_in_502
 
 router = APIRouter(dependencies=[Depends(check_operations_account)])
